+++ packages/runtime/src/index.ts
@@ -3,6 +3,7 @@ export * from "./dependency-graph";
 export * from "./function-registry";
 export * from "./manager-adapter";
 export * from "./orchestrator";
+export * from "./ws-events";
 export * from "./rule-engine";
 export * from "./template-resolver";
 export * from "./types";